% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whensJuly4.R
\name{whensJuly4}
\alias{whensJuly4}
\title{Days Until Birthday}
\usage{
whensJuly4(date = NULL)
}
\arguments{
\item{date}{(Character) The target date in "YYYY-MM-DD" format. Defaults to the current date.}
}
\value{
Returns a message with number of days left until July 4th.
}
\description{
A simple function to determine how many days are left until July 4th.
}
\details{
The description says it all!
}
\examples{
library(whensJuly4)
Sys.Date() # for reference
whensJuly4()
}
\author{
Walid Nashashibi (\url{https://github.com/wnashash/})
}
